Update burgers.js
I don't know what I was thinking!
I totally forgot that using the arrow functions in the JSX syntax cause the value of $(this) or 'this' to refer to the window object that it is called in and not the element that calls it. You just need to change your function syntax.